requirements.txt
pandas==2.2.3
pdf2image==1.17.0
pdfplumber==0.11.4
PyMuPDF==1.24.11
PyPDF2==3.0.1
pytesseract==0.3.13
pytest==8.3.3
python-docx==1.1.2
python-dotenv==1.0.1
python-pptx==1.0.2
pillow==10.4.0
msoffcrypto-tool==5.4.2

# Python Standard Library Versions
# os: 3.12.3
# typing: 3.12.3
# abc: 3.12.3
# json: 3.12.3
# io: 3.12.3
# sqlite3: 3.12.3